+++ civis/cli/_cli_commands.py
@@ -3,8 +3,10 @@
 """
 Additional commands to add to the CLI beyond the OpenAPI spec.
 """
-
+from __future__ import print_function
+import functools
 import os
+import sys
 
 import click
 import requests
@@ -68,6 +70,90 @@ def files_download_cmd(file_id, path):
     """Download a Civis File to a specified local path."""
     with open(path, 'wb') as f:
         civis_to_file(file_id, f)
+
+
+@click.command('sql')
+@click.option('--dbname', '-d', type=str, required=True,
+              help='Execute the query on this Civis Platform database')
+@click.option('--command', '-c', type=str, default=None,
+              help='Execute a single input command string')
+@click.option('--filename', '-f', type=click.Path(exists=True),
+              help='Execute a query read from the given file')
+@click.option('--output', '-o', type=click.Path(),
+              help='Download query results to this file')
+@click.option('--quiet', '-q', is_flag=True, help='Suppress screen output')
+@click.option('-n', type=int, default=100,
+              help="Display up to this many rows of the result. Max 100.")
+def sql_cmd(dbname, command, filename, output, quiet, n):
+    """\b Execute a SQL query in Civis Platform
+
+    If neither a command nor an input file is specified, read
+    the SQL command from stdin.
+    If writing to an output file, use a Civis SQL script and write the
+    entire query output to the specified file.
+    If not writing to an output file, use a Civis Query, and return a
+    preview of the results, up to a maximum of 100 rows.
+    """
+    if filename:
+        with open(filename, 'rt') as f:
+            sql = f.read()
+    elif not command:
+        # Read the SQL query from user input. This also allows use of a heredoc
+        lines = []
+        while True:
+            try:
+                _i = input()
+            except (KeyboardInterrupt, EOFError):
+                # The end of a heredoc produces an EOFError.
+                break
+            if not _i:
+                break
+            else:
+                lines.append(_i)
+        sql = '\n'.join(lines)
+    else:
+        sql = command
+
+    if not sql:
+        # If the user didn't enter a query, exit.
+        if not quiet:
+            print('Did not receive a SQL query.', file=sys.stderr)
+        return
+
+    if not quiet:
+        print('\nExecuting query...', file=sys.stderr)
+    if output:
+        fut = civis.io.civis_to_csv(output, sql, database=dbname)
+        fut.result()  # Block for completion and raise exceptions if any
+        if not quiet:
+            print("Downloaded the result of the query to %s." % output,
+                  file=sys.stderr)
+    else:
+        fut = civis.io.query_civis(sql, database=dbname,
+                                   preview_rows=n, polling_interval=3)
+        cols = fut.result()['result_columns']
+        rows = fut.result()['result_rows']
+        if not quiet:
+            print('...Query complete.\n', file=sys.stderr)
+            print(_str_table_result(cols, rows))
+
+
+def _str_table_result(cols, rows):
+    """Turn a Civis Query result into a readable table."""
+    # Determine the maximum width of each column.
+    # First find the width of each element in each row, then find the max
+    # width in each position.
+    max_len = functools.reduce(
+        lambda x, y: [max(z) for z in zip(x, y)],
+        [[len(_v) for _v in _r] for _r in [cols] + rows])
+
+    header_str = " | ".join("{0:<{width}}".format(_v, width=_l)
+                            for _l, _v in zip(max_len, cols))
+    tb_strs = [header_str, len(header_str) * '-']
+    for row in rows:
+        tb_strs.append(" | ".join("{0:>{width}}".format(_v, width=_l)
+                                  for _l, _v in zip(max_len, row)))
+    return '\n'.join(tb_strs)
 
 
 @click.command('download')